refactor(whispering): collapse dead accelerator code to the migration parser
After the rdev swap, the only reachable export of utils/accelerator.ts was
acceleratorToKeyBinding (one caller: the device-config migration). The
Electron-accelerator machinery it sat in, pressedKeysToAccelerator,
isValidElectronAccelerator, the Accelerator brand, the error types, the
modifier-priority sort, was all dead. Collapse to just the migration parser
and rename the file to legacy-accelerator.ts so the name matches its job.
